refactor: privatize typing module
