+++ falcon/asgi/request.py
@@ -34,13 +34,13 @@ from typing import (
 from falcon import errors
 from falcon import request
 from falcon import request_helpers as helpers
+from falcon._typing import AsgiReceive
+from falcon._typing import MISSING
+from falcon._typing import MissingOr
+from falcon._typing import StoreArgument
 from falcon.asgi_spec import AsgiEvent
 from falcon.constants import SINGLETON_HEADERS
 from falcon.forwarded import Forwarded
-from falcon.typing import AsgiReceive
-from falcon.typing import MISSING
-from falcon.typing import MissingOr
-from falcon.typing import StoreArgument
 from falcon.util import deprecation
 from falcon.util import ETag
 from falcon.util.uri import parse_host
