+++ falcon/media/handlers.py
@@ -18,6 +18,8 @@ from typing import (
 )
 
 from falcon import errors
+from falcon._typing import DeserializeSync
+from falcon._typing import SerializeSync
 from falcon.constants import MEDIA_JSON
 from falcon.constants import MEDIA_MULTIPART
 from falcon.constants import MEDIA_URLENCODED
@@ -28,8 +30,6 @@ from falcon.media.json import JSONHandler
 from falcon.media.multipart import MultipartFormHandler
 from falcon.media.multipart import MultipartParseOptions
 from falcon.media.urlencoded import URLEncodedFormHandler
-from falcon.typing import DeserializeSync
-from falcon.typing import SerializeSync
 from falcon.util import misc
 from falcon.vendor import mimeparse
