+++ falcon/routing/compiled.py
@@ -35,10 +35,10 @@ from typing import (
     Union,
 )
 
+from falcon._typing import MethodDict
 from falcon.routing import converters
 from falcon.routing.util import map_http_methods
 from falcon.routing.util import set_default_responders
-from falcon.typing import MethodDict
 from falcon.util.misc import is_python_func
 from falcon.util.sync import _should_wrap_non_coroutines
 from falcon.util.sync import wrap_sync_to_async
